% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebcd.R
\name{ebcd}
\alias{ebcd}
\title{Empirical Bayes Covariance Decomposition}
\usage{
ebcd(
  X = NULL,
  S = NULL,
  C = NULL,
  N = NULL,
  Kmax = 5,
  ebnm_fn = ebnm::ebnm_point_laplace,
  tol_greedy = 1e-06,
  maxiter_greedy = 500,
  tol_backfit = 1e-06,
  maxiter_backfit = 5000
)
}
\arguments{
\item{X}{A N-by-P data matrix with N observations and P variables}

\item{S}{A P-by-P covariance/Gram matrix \eqn{=X'X/N}}

\item{C}{A P-by-P compact target matrix such that \eqn{C'C=X'X}}

\item{N}{The number of observations}

\item{Kmax}{The maximum number of factors to be added.}

\item{ebnm_fn}{The ebnm function to be used.}

\item{tol_greedy}{The convergence tolerance parameter when optimizing the greedily added factor.}

\item{maxiter_greedy}{The maximum number of iterations when optimizing the greedily added factor.}

\item{tol_backfit}{The convergence tolerance parameter for the backfit procedure.}

\item{maxiter_backfit}{The maximum number of iterations for the backfit procedure.}
}
\value{
An ebcd object.
}
\description{
Empirical Bayes Covariance Decomposition
}
