% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebcd_backfit.R
\name{ebcd_backfit}
\alias{ebcd_backfit}
\title{Title}
\usage{
ebcd_backfit(ebcd, tol = 1e-06, maxiter = 5000)
}
\arguments{
\item{ebcd}{An ebcd object.}

\item{tol}{The convergence tolerance parameter.}

\item{maxiter}{The maximum number of iterations for the backfit procedure.}
}
\value{
An ebcd object.
}
\description{
Title
}
