% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebcd_greedy.R
\name{ebcd_greedy}
\alias{ebcd_greedy}
\title{Greedily add factors to an ebcd object}
\usage{
ebcd_greedy(
  ebcd_obj,
  Kmax = 1,
  ebnm_fn = ebnm::ebnm_point_laplace,
  tol = 1e-06,
  maxiter = 500
)
}
\arguments{
\item{ebcd_obj}{An ebcd object.}

\item{Kmax}{The maximum number of factors to be added.}

\item{ebnm_fn}{The ebnm function to be used.}

\item{tol}{The convergence tolerance parameter.}

\item{maxiter}{The maximum number of iterations when optimizing the greedily added factor.}
}
\value{
An ebcd object.
}
\description{
Greedily add factors to an ebcd object
}
