% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_means_loglik.R
\name{normal_means_loglik}
\alias{normal_means_loglik}
\title{Compute the Normal means log likelihood}
\usage{
normal_means_loglik(x, s, Et, Et2)
}
\arguments{
\item{x}{A vector of observations in an EBNM problem}

\item{s}{A vector of standard errors in an EBNM problem}

\item{Et}{A vector of posterior means}

\item{Et2}{A vector of posterior mean of second moments}
}
\description{
Compute the Normal means log likelihood
}
\references{
This function is copied from flashier:::normal.means.loglik
(https://github.com/willwerscheid/flashier/blob/master/R/objective.R)
}
