% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_means_loglik.R
\name{normal_means_loglik}
\alias{normal_means_loglik}
\title{Compute the Normal means log likelihood}
\usage{
normal_means_loglik(x, s, Et, Et2)
}
\description{
Compute the Normal means log likelihood
}
\references{
This function is copied from flashier:::normal.means.loglik
(https://github.com/willwerscheid/flashier/blob/master/R/objective.R)
}
